+++ src/app/wish/(list)/stores/page.tsx
@@ -1,5 +1,5 @@
 import WishStroeList from '@/blocks/wish/(list)/stores/WishStoreList';
-import { storeQueryKey } from '@/domains/store/queries/queryKey';
+import { storeQueryKey } from '@/shared/queries/queryKey';
 import wishService from '@/domains/wish/queries/service';
 import { INITIAL_CURSOR } from '@/shared/constants/cursor';
 import { HydrationBoundary, QueryClient, dehydrate } from '@tanstack/react-query';
