--- src/domains/store/queries/queryKey.ts
@@ -1,5 +1,0 @@
-export const storeQueryKey = {
-  all: ['store'],
-  lists: () => [...storeQueryKey.all, 'list'],
-  list: (filter: string) => [...storeQueryKey.lists(), filter]
-};
