+++ src/domains/wish/queries/useAddWishStoreMutation.ts
@@ -2,7 +2,7 @@ import useToastNewVer from '@/shared/hooks/useToastNewVer';
 import { InfiniteData, useMutation, useQueryClient } from '@tanstack/react-query';
 import { Cursor } from '@/shared/types/response';
 import { IStoreType } from '@/domains/store/types/store';
-import { storeQueryKey } from '@/domains/store/queries/queryKey';
+import { storeQueryKey } from '@/shared/queries/queryKey';
 import { updateInfiniteQueryCache } from '../../../shared/utils/queryCache';
 import wishService from './service';
